chore(chat): resolve HMR issues
