+++ apps/chat/src/utils/app/conversation.ts
@@ -393,3 +393,25 @@ export const isLoadedConversationEntity = (
   entity: ShareEntity,
 ): entity is Conversation =>
   isConversationInfoEntity(entity) && entity.status === UploadStatus.LOADED;
+
+export function getMessageCustomContent(
+  message: Message,
+): Partial<Message> | undefined {
+  return message.custom_content?.state ||
+    message.custom_content?.attachments?.length ||
+    message.custom_content?.form_value ||
+    message.custom_content?.form_schema
+    ? {
+        custom_content: {
+          attachments:
+            message.role !== Role.Assistant &&
+            message.custom_content?.attachments?.length
+              ? message.custom_content?.attachments
+              : undefined,
+          state: message.custom_content?.state,
+          form_value: message.custom_content?.form_value,
+          form_schema: message.custom_content?.form_schema,
+        },
+      }
+    : undefined;
+}
